Moved logic to getPossibleCustomPostTypes
