+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/WPDataModel/WPDataModelProviderInterface.php
@@ -6,4 +6,8 @@ namespace GraphQLAPI\GraphQLAPI\WPDataModel;
 
 interface WPDataModelProviderInterface
 {
+    /**
+     * @return string[]
+     */
+    public function getPossibleCustomPostTypes(): array;
 }
